style: Norminette demands
