+++ include/parse.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parse.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gasouza <gasouza@student.42sp.org.br>      +#+  +:+       +#+        */
+/*   By: rmiranda <rmiranda@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/06 16:03:07 by rmiranda          #+#    #+#             */
-/*   Updated: 2023/10/22 14:33:45 by gasouza          ###   ########.fr       */
+/*   Updated: 2023/10/24 00:04:36 by rmiranda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,17 +20,16 @@
 # define KEEP_NEWLINE 0
 # define C 1
 # define F 0
+# define ALLOWED_CHARACTERS " 01NWSE"
 
 // Getter
-int	add_next_instruction(char *line, t_map_info *info_ptr);
-int	add_line_to_map(t_map_info *info_ptr, char *line);
+int		add_next_instruction(char *line, t_map_info *info_ptr);
 
 // Validation
 int		map_validation(t_map_info *info_ptr);
+int		add_line_to_map(t_map_info *info_ptr, char *line);
 char	**normalize_map(t_map_info *info_ptr);
 int		assert_is_not_surrounded_by_wall(t_map_info info);
-int		assert_is_exposed_floor(t_map_info info, int xx, int yy);
-int		bounderies_check_horizontal_axis(t_map_info info, int xx, int yy);
-int		bounderies_check_vertical_axis(t_map_info info, int xx, int yy);
+void	space_injest_func(unsigned int nb, char *c);
 
 #endif
